Ajustes de arquitetura
