+++ app/src/main/java/com/arctouch/primeiroapp/DetailsScreen.kt
@@ -19,6 +19,7 @@ import androidx.compose.ui.text.style.TextIndent
 import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.arctouch.primeiroapp.models.Filme
 
 
 @Composable
