+++ app/src/main/java/com/arctouch/primeiroapp/FilmesUiState.kt
@@ -1,5 +1,7 @@
 package com.arctouch.primeiroapp
 
+import com.arctouch.primeiroapp.models.Filme
+
 sealed interface FilmesUiState {
     data object Carregando : FilmesUiState
     data class Carregado(val filmes: List<Filme>) : FilmesUiState
