+++ app/src/main/java/com/arctouch/primeiroapp/FilmesViewModel.kt
@@ -2,6 +2,7 @@ package com.arctouch.primeiroapp
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.arctouch.primeiroapp.repository.FilmesRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
