--- app/src/main/java/com/arctouch/primeiroapp/repository/FilmesUiState.kt
@@ -1,4 +1,0 @@
-package com.arctouch.primeiroapp.repository
-
-class FilmesUiState {
-}
