+++ app/src/test/java/com/arctouch/primeiroapp/PerfilViewModelTest.kt
@@ -1,6 +1,6 @@
 package com.arctouch.primeiroapp
 
-import com.arctouch.primeiroapp.repository.PerfilData
+import com.arctouch.primeiroapp.models.PerfilData
 import com.arctouch.primeiroapp.repository.PerfilRepository
 import io.mockk.confirmVerified
 import io.mockk.every
